requirements.txt
python-jobspy>=1.1.80
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
gunicorn>=22.0.0
pandas>=2.0.0
httpx>=0.27.0